Bit 'o flesh for the bindnode impl.
This has been sitting on my hard drive quite a while already, so I
decided to clean it up juuuust enough that it compiles so I can push
it.  But as you can see, there's a LOT of TODOs dangling here.

I'm not at all still sure those comments about "root node" in the
interface are going to stand up to scrutiny over time.  There's
definitely going to be some distinction between "Node we're traversing"
and "Node we are putting in a serialized block of bytes and can
think it's reasonable to address by hash", but figuring out the best
ergonomics of that is probably going to take a while.

I'm going to start on another simpler just-backed-by-maps-n-stuff Node
impl before going deeper with this one, because I suspect it'll make
tests easier to write, which will then make this whole pile of
reflection easier to test down the road as well.

Signed-off-by: Eric Myhre <hash@exultant.us>