Add cid lib and deps.
Note: I am not very pleased with how many deps this core and simple
data type has.  These are basically all for the hashing, and very, very
little actual hashing code at that, and I really wish the rule about
"a little vendoring is better than a little dependency" would be
applied here -- pulling in x/crypto and x/sys so that I can use a
type that's *a wrapper around string* is thoroughly ridiculous.

Signed-off-by: Eric Myhre <hash@exultant.us>