Documentation aimed at overview for developers.
Contents:

- A few high level things, mostly just to give enough minimal context and
terminology to continue, and shelling out to the ipld/specs repo for
details.

- "Everything is a Node"

- Code layout with brief words about the intent behind each tree of
packages.

One of the main things I wanted to get on paper this morning is the
distinction between the reified/active/usable type-checker types,
versus the tree that represents them serially.  The former may contain
cyclic references; the latter obviously cannot.  We don't yet have
separate packages and types for these, but should asap.  This
description all snuck in under "code layout".

Signed-off-by: Eric Myhre <hash@exultant.us>