Impl and tests for fluent.Recover.
Used the ipldfree.Node impl concretely to make the tests work.
In theory any impl should of course work, but it's not useful or
interesting to do a compatibility matrix here -- particularly given
that we use the fluent system in the compataibility tests for all
the other Node impl packages.

Signed-off-by: Eric Myhre <hash@exultant.us>