Fix semantic bug in TypeUnion: lookup key is not always the member type's name!
While fixing this it occurs to me that we could in theory also support
more than one name alias that will be mapped onto the same concrete type.
...
But Let's Not with that.  It would increase the code complexity and more
importantly, *it would be lossy to do so*.  So let's not open that can
of worms: straying into lossiness territory is far worse than whatever
other hypothetical problems one might imagine trying to solve.

Signed-off-by: Eric Myhre <hash@exultant.us>