Introduce typedeclaration package.  (Think "AST".)
All of these struct definitions are horrifying similar in construction
to the same-named definitions in the typesystem package: simply ever
instance of a Type pointer is instead a TypeName string.  Despite being
unattractive, this is correct: since typesystem can have pointers and
cycles, and typedeclaration can't, this is what we end up with.

Signed-off-by: Eric Myhre <hash@exultant.us>