Introduce ipld.NodeBuilder interface.
This will replace, deprecate, annihilate, obviate, oblivate, and
generally nuke the MutableNode interface.  MutableNode was a bad, bad
idea and made many things terribly complicated.  Immutable interfaces
are shaping up much, much better and I'm excited to move NodeBuidler
forward.

I tried for some time to figure out a useful way to do a builder
pattern in the fluent package, which would work *upon* the MutableNode
interface... but all attempts to do that ended poorly with code that
was both messy internally and complex and error prone for callers.
I've given up all such attempts.

It also seems to be true that we should have NodeBuilder *instances*
which may be per Node.  This wasn't obvious at first, but I think it'll
let us carry a lot of important information in a low-fuss way, and it
solves a *lot* of problems that cropped up when trying to implement
larger scale generic transformation methods on Node graphs (namely, the
frequent need to be able to create a new Node in the middle of a
process; and handing down a factory method to each of those points was
horrifically verbose and caused complexity that was not essential).
(Unfortunately, I can't really point at examples of *how* bad this got,
because I've never committed any of those research spikes; fortunately,
no one else will now ever have to *see* that mess, since we've got
better solutions figured out.)

MutableNode will go away in the next handful of commits, but I wanted
to land the new one first in a smaller diff.  Removing MutableNode
may be a fair amount of messy work, because of how it's used in nearly
all of our test fixture setup to date!  Uffdah.

(Node will get a GetBuilder method (or something of similar intent but
better name) at the same time we get rid of MutableNode; I'm eliding
it from this commit for the same reasons about diff size avoidance.)

Signed-off-by: Eric Myhre <hash@exultant.us>