Additional doc about carry powers of NodeBuilder.
Specifically in regards to typed nodes, which are the main consumers of
this that I imagine right now (though we may find others in the
future).

Typed nodes which are implemented by *native* concrete types (e.g.
presumably by codegen -- though we haven't implemented any of this yet)
will *especially* hit this detail: they'll return builders that produce
new instances of their own concrete native type!

And all of the build methods that actually produce a new Node now also
have the option to return an error.  This follows from the docs;
typed nodes actually can and will reject some attempts at invalidly-
typed creation!

We'll probably want fluent wrappers of NodeBuilder as well, as a result
of these error returns.  I'll defer that until it's forced.
Lazy/just-in-time evaluation of the programmer :P

Signed-off-by: Eric Myhre <hash@exultant.us>