First pass on Transform methods.
The docs in the diff body do most of the explaining work here.

I'm not sure if these should be in the fluent package or not.
Probably it will be wiser to have a non-fluent variant first, and wrap
those in the fluent package.  I wrote them here first, though, so,
here they stay for this commit.

I'm checkpointing this here, but there's at least one major issue I've
already realized: with these interfaces, we can't keep correct paths
when we do *nested* Transform calls.  Perhaps that's survivable; but I
find it very unattractive.  So, we'll do some additional work to hammer
that back out -- the tricky part will just be doing it without
increasing the amount of complexity that's user-facing.

Also -- yes, I know the parent commit says "we'll get rid of
MutableNode" next -- but it's basically this Transform stuff that's
been the driving force behind most of the other traversal and builder
work going on today, so...  We'll nudge both along in tandem for a bit.

Signed-off-by: Eric Myhre <hash@exultant.us>