UnionRepresentation... union.  yes.
This is actually a fairly big improvement on what we'd written for the
internal implementation in Go.  The representation *should* be a union,
in general.  We'll do something similar for structs, and everything
that has multiple options for representation strategy.

One FIXME bugaboo: it's now absolutely unclear what to put as keys in
the members map, as although there is a key, which one to use depends
on the representation kind.  It would feel odd to not have the
membership be clear before reading through to the representation info,
but I guess that is defacto where that information lives, so...

Signed-off-by: Eric Myhre <hash@exultant.us>