Define TypeStruct and StructField.
StructRepresentation also offers at least two options; and in the
future, we might anticipate adding more options as well (suppose for
example some kind of struct-to-string conversion systems; of which
their could be many, but imagine for example one like matching a
string like the "options" in your /etc/fstab: simple comma separated
keys and values should be definable enough, no?).

And now we have to wrangle the TypeTerm recursivity to completeness:
how shall we concisely and correctly represent the declaration of
inline declarations of map and array kinded types?  I've left this
a big honking TODO for this commit; it's touchy.

Signed-off-by: Eric Myhre <hash@exultant.us>