Clean up typed/declaration/* types; closer to back in line with schema.
Removed the 'name' field from all of them.  Doesn't match schema.
While it's true that we'll want a name property on the reflective
typesystem.Type objects, that's... well, that's over there.

I'm starting to make fairly wild zigzags through here; with reason:
I'm trying to get as quickly as possible to a bootstrap of the essenses
of codegen.  The hope is to get there fast, and power it with a few
hardcoded values; then get that to churn out a replacement for the
hand-written declarations package types; and then teach the codegen
to write more and more of the e.g. ipld.Node interfaces and then
the serialization bridges and then the NodeBuilder stuff, etc.
The core coverage for getting the codegen to go, though, I think it's
actually sufficiently smaller than all those parts that maybe this
will be reasonable to pull off this way.  Time will tell.  (Shortly!)

Dropping the typed/system/construct.go file temporarily, as the changes
to the other typedeclaration files has broken it, and maintaining it
through the shaking seems like it will generate inordinately more
work than it's worth.  It'll be revived soon; it's mostly on the right
track, it's just not on the critical path atm.

Signed-off-by: Eric Myhre <hash@exultant.us>