Further schema-schema fixes for 'default' values.
Continues the work started in 2263295, with the purpose of fixing
undesired three-valued logic by disambiguating 'optional' from
'default'.  Some comments in the schema now address this in detail,
including description of the membership cardinality.

Some incongruities in the JSON have been fixed;
and a significant new set of definitions for struct-represented-as-map
configuration details is introduced.

I gave some brief consideration to eliding StructRepresentation_Map
and instead simply using a typedef'd map, and it would be possible
since it only has one field... however, I'm opting to keep it because
I'm willing to bet we'll come up with more properties someday which
are for the whole struct rather than just per field, and thus it's
unlikely that StructRepresentation_Map will stay single-field forever.

The 'default' value itself in the new
StructRepresentation_Map_FieldDetails type is interesting: it's the
first time we've had an excuse to use an "Any" cop-out.
And even more interestingly, it's almost certainly invalid: as
commented inline in the diff, that field would probably be even better
represented as a kinded union which makes it clear that only the scalar
primitives are welcome here.
We'll almost certainly introduce that union in a future commit, but
I'm gonna let that thought marinate for a while first.
There are related interesting questions about how we want to handle
the DSL parsing in this area, and those might be good to think about
more deeply before finalizing this part of the schema.

Signed-off-by: Eric Myhre <hash@exultant.us>