fluent.Node.Keys should panic if errors on route.
Like all other methods on fluent.Node that panic the instant a concrete
type of anything is asserted outside of traversal attempts.

There's not much sense in holding off until trying to start an
iteration that's most certainly going to fail; and we do assume that
anyone making the iterator intends to use it.

Signed-off-by: Eric Myhre <hash@exultant.us>