Updated Traversal; introduce TraversalProgress.
The policy on partial progress results is now explicitly "no".
It wasn't useful.
Tests have been updated accordingly.  (The relevant tests *do* still
still assert that the traversal errors happen where they should; it's
just via the... well, the error value.  Sense, neh?)

We're introducing a TraversalProgress struct.  This is roughly
following "Option 2" from the earlier "this composition is imperfect!"
lamentation block: it allows us to keep carrying Path from existing
moves across the graph at all times.

(I suspect we're going to incur somewhat unfortunate memory allocation
thrash in large-scale operation because of all these immutable Path
objects, which (ultra)unfortunately can't even reuse the n+1 length
arrays when their prior sibling already allocated them... However,
I care a lot more about getting the composability and ergonomics of
these interfaces right first; and we can discuss how to hack shortcuts
back in later.  (Perhaps the more visit-oriented functions will do
some fancy footwork, and combine it with warnings that the Path objects
handed the visitor are only valid for its scope.  This is sharp-edged,
but seems to be common in practice in this kind of library; basically
every performant git library I've ever seen has caveats like this on
its iterators, for example.  Anyway!  Not doing that yet.))

Context is coming along for the ride on TraversalProgress!
This is discussed in another one of the large comment blocks that's
disappearing in this commit.

Newly exported Path.Join method.  Does what it says on the tin.

Usage of this whole new TraversalProgress system is limited: we're just
cutting away the incomplete cruft of first-draft functions in this
diff whenever they got in the way, and newly designed
Traverse/Visit/Transform/etc functions which correctly use
TraversalProgress will be coming up in future diffs.

Signed-off-by: Eric Myhre <hash@exultant.us>