Moremore schema migration prose.
I think this is as much as I can write about this for now.

Regarding that last bit about not having total migration magic:
I'd certainly be neato to offer more auto-migration tools, based on
perhaps a "patch"ing approach as outlined in
https://github.com/ipld/js-ipld/issues/66#issuecomment-266345127 ,
or on generalized recursion schemes, or a combination.
However... that's a tad downstream of the present ;)

Signed-off-by: Eric Myhre <hash@exultant.us>